DW-2379 Delegated to validator from decryption processor
